More utility factoring and inproved text output
