+++ AStarAlgorithmsProject/AStarAlgorithmsProject.csproj
@@ -61,6 +61,7 @@
     <Compile Include="Alg Scripts\BreadthFirstSolver.cs" />
     <Compile Include="Alg Scripts\Point.cs" />
     <Compile Include="Alg Scripts\SolverUtils.cs" />
+    <Compile Include="Alg Scripts\Tile.cs" />
     <Compile Include="App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
