+++ AStarAlgorithmsProject/Alg Scripts/SolverUtils.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace AStarAlgorithmsProject
 {
-    enum TileStates { Unchecked, Open, Closed }; // enum to title and track the 3 states a tile can be in.
+    public enum TileStates { Unchecked, Open, Closed }; // enum for 3 states a tile can be in.
 
     static class SolverUtils
     {
@@ -14,17 +14,15 @@ namespace AStarAlgorithmsProject
         public static Point[] GetAdvancedDirections()
         {
             return new Point[] {
-            new Point(-1, 1), new Point(0, 1), new Point(1, 1),  new Point(-1, 0),
-            new Point(1, 0),   new Point(-1, -1), new Point(0, -1), new Point(1, -1)
-                                                };
+                new Point(-1, 1), new Point(0, 1), new Point(1, 1),  new Point(-1, 0),
+                new Point(1, 0),   new Point(-1, -1), new Point(0, -1), new Point(1, -1)};
         }
 
         // in order: N, W, E, S
         public static Point[] GetSimpleDirections()
         {
             return new Point[] {
-            new Point(0, 1), new Point(-1, 0), new Point(1, 0), new Point(0, -1)
-                                                };
+                new Point(0, 1), new Point(-1, 0), new Point(1, 0), new Point(0, -1)};
         }
     }
 }
